Task-18: Pass parameters dynamically to powershell file

+++ src/DiarizationInterpreter.java
@@ -12,7 +12,7 @@ public class DiarizationInterpreter extends Interpreter {
         jsonObject = readJsonFile(path);
         diarizationObject = (JSONObject) jsonObject.get("Diarization");
 
-        this.duration = (long) diarizationObject.get("duration");
+        this.duration = (double) diarizationObject.get("duration");
         this.segmentStart = (ArrayList<Double>) diarizationObject.get("segmentStart");
         this.label = (ArrayList<String>) diarizationObject.get("label");
         this.segmentEnd = (ArrayList<String>) diarizationObject.get("segmentEnd");
